一部の関数を別のファイルに退避
