+++ src/rmmz/plugin/schema/compile/kinds/core/parse/state.ts
@@ -1,5 +1,6 @@
 import { addBasePlugin, addOrderAfter, addOrderBefore } from "./dependencies";
-import type { ParseState } from "./internalTypes";
+import type { OptionsState, ParseState } from "./internalTypes";
+import { addOption, addValue } from "./option";
 
 export const handleBase = (state: ParseState, value: string): ParseState => {
   return { ...state, dependencies: addBasePlugin(state.dependencies, value) };
@@ -22,5 +23,32 @@ export const handleOrderBefore = (
   return {
     ...state,
     dependencies: addOrderBefore(state.dependencies, value),
+  };
+};
+
+export const handleOption = (state: ParseState, value: string): ParseState => {
+  if (!state.currentParam) {
+    return state;
+  }
+
+  const newOption: OptionsState = addOption(
+    state.currentOption ?? { items: [] },
+    value
+  );
+
+  return {
+    ...state,
+    currentOption: newOption,
+  };
+};
+
+export const handleValue = (state: ParseState, value: string): ParseState => {
+  if (!state.currentOption) {
+    return state;
+  }
+  const newOption: OptionsState = addValue(state.currentOption, value);
+  return {
+    ...state,
+    currentOption: newOption,
   };
 };
